+++ Zenith/Models/SaleItem.cs
@@ -33,10 +33,10 @@ namespace Zenith.Models
         public virtual Material Material { get; set; }
 
         [Reactive]
-        public long UnitPrice { get; set; }
+        public float UnitPrice { get; set; }
 
         [Reactive]
-        public int Count { get; set; }
+        public float Count { get; set; }
 
         [Reactive]
         public CountUnits SaleCountUnit { get; set; }
@@ -47,7 +47,7 @@ namespace Zenith.Models
 
         [NotMapped]
         [Reactive]
-        public long TotalPrice { get; set; }
+        public float TotalPrice { get; set; }
 
         [Reactive]
         public string Comment { get; set; } = string.Empty;
